+++ src/config.toml
@@ -1,3 +1,31 @@
+# [[drone]]
+# id = 0
+# connected_node_ids = [1, 2]
+# pdr = 0.05
+#
+# [[drone]]
+# id = 1
+# connected_node_ids = [0,2]
+# pdr = 0.03
+#
+# [[drone]]
+# id = 2
+# connected_node_ids = [1,0]
+# pdr = 0.14
+#
+# [[client]]
+# id = 3
+# connected_drone_ids = [3, 2]
+#
+# [[client]]
+# id = 4
+# connected_drone_ids = [1]
+#
+#
+# [[server]]
+# id = 5
+# connected_drone_ids = [2,3]
+
 [[drone]]
 id = 0
 connected_node_ids = [1, 2]
@@ -5,23 +33,38 @@ pdr = 0.05
 
 [[drone]]
 id = 1
-connected_node_ids = [0,2]
-pdr = 0.03
+connected_node_ids = [0, 2, 5]
+pdr = 0.08
 
 [[drone]]
 id = 2
-connected_node_ids = [1,0]
-pdr = 0.14
+connected_node_ids = [0, 1, 3, 4, 5]
+pdr = 0.02
 
-[[client]]
+[[drone]]
 id = 3
-connected_drone_ids = [3, 2]
+connected_node_ids = [2]
+pdr = 0.15
+
+[[drone]]
+id = 4
+connected_node_ids = [2, 5]
+pdr = 0.15
+
+[[drone]]
+id = 5
+connected_node_ids = [1,2,4]
+pdr = 0.11
 
 [[client]]
-id = 4
-connected_drone_ids = [1]
+id = 6
+connected_drone_ids = [0]
+
+[[server]]
+id = 7
+connected_drone_ids = [1,5]
 
 
 [[server]]
-id = 5
-connected_drone_ids = [2,3]
+id = 8
+connected_drone_ids = [3,4]
