+++ src/tests/mod.rs
@@ -1,2 +1,2 @@
 pub mod send_msg;
-pub mod send_query;
+pub mod send_flood_request;
